refactor: move previous page components out of /app
