+++ frontend/src/components/pages/auth/Signup.tsx
@@ -81,7 +81,7 @@ const Signup = () => {
         <div className="mt-4 text-center">
           <p className="text-sm">
             Already have an account?{" "}
-            <a href="/login" className="text-blue-600 hover:underline">
+            <a href="/auth/Login" className="text-blue-600 hover:underline">
               Login
             </a>
           </p>
