+++ frontend/src/components/pages/event/Schedule.tsx
